Separate kit strengths by team
Rather than having one set of variables,
have a set per-team.

Also, use the inverse logit function to calculate
probability. This should be slightly faster.

And explicitly model the fact that alpha does
not win 50% of the time. Make it so that evenly
matched teams win at the base rate for alpha.